Move stack, merge and operation as submodules of eval
