+++ src/eval/operation.rs
@@ -10,11 +10,11 @@ use crate::error::EvalError;
 use crate::eval::{subst, CallStack, Closure, Environment};
 use crate::identifier::Ident;
 use crate::label::ty_path;
-use crate::merge;
-use crate::merge::{merge, MergeMode};
+use super::merge;
+use super::merge::{merge, MergeMode};
 use crate::mk_record;
 use crate::position::TermPos;
-use crate::stack::Stack;
+use super::stack::Stack;
 use crate::term::make as mk_term;
 use crate::term::{BinaryOp, NAryOp, RichTerm, StrChunk, Term, UnaryOp};
 use crate::transform::Closurizable;
