+++ src/lib.rs
@@ -5,14 +5,11 @@ pub mod error;
 pub mod eval;
 pub mod identifier;
 pub mod label;
-pub mod merge;
-pub mod operation;
 pub mod parser;
 pub mod position;
 pub mod program;
 pub mod repl;
 pub mod serialize;
-pub mod stack;
 pub mod stdlib;
 pub mod term;
 pub mod transform;
